Crate penumbra_sdk_view

The view RPC library for the Penumbra Zone.

This crate provides a ViewClient and a ViewServer. These form a client-server pair to synchronize and interact with public chain state using one or more full viewing keys. See the documentation of ViewClient and a ViewServer for more information.

This crate also provides a Planner. This is a planner for TransactionPlan.

Finally, this crate provides a Storage type for managing persistent sqlite storage.

Structs

Planner

A planner for a TransactionPlan that can fill in the required spends and change outputs upon finalization to make a transaction balance.

SpendableNoteRecord

Corresponds to the SpendableNoteRecord proto

StatusStreamResponse

Storage

SwapRecord

TransactionInfo

ViewServer

A service that synchronizes private chain state and responds to queries about it.

Traits

ViewClient

The view protocol is used by a view client, who wants to do some transaction-related actions, to request data from a view service, which is responsible for synchronizing and scanning the public chain state with one or more full viewing keys.

Functions

register_metrics

Registers all metrics used by this crate.